Secured edit and delete idea endpoints

+++ app/views.py
@@ -149,6 +149,8 @@ def create_idea():
 def edit_idea(idea_id):
     if not current_user.is_authenticated:
         return redirect(url_for('login'))
+    if current_user.id != get_idea(idea_id).author.id:
+        abort(403)
     if idea_exists(idea_id):
         idea_to_edit = get_idea(idea_id)
         form = EditIdeaForm(title=idea_to_edit.title,
@@ -169,6 +171,8 @@ def edit_idea(idea_id):
 def delete_idea(idea_id):
     if not current_user.is_authenticated:
         return redirect(url_for('login'))
+    if current_user.id != get_idea(idea_id).author.id:
+        abort(403)
     delete_idea_by_id(idea_id)
     flash('Your idea has been deleted!', 'info')
     return redirect(url_for('profile'))
